0.3.0 version
style updates to circuit grid, trying to make it look good on mobile

entirely REMOVED the single event template from this plugin as it was not needed

REMOVED publish/events directory in favor of the new upsert handler which does the same thing but also handles updating events that may have changed data some
