+++ inc/blocks/EventDetails/block.json
@@ -1,7 +1,7 @@
 {
     "apiVersion": 2,
     "name": "datamachine-events/event-details",
-    "version": "0.2.2",
+    "version": "0.2.4",
     "title": "Event Details",
     "category": "datamachine-events",
     "icon": "calendar-alt",
