+++ inc/blocks/calendar/DisplayStyles/CircuitGrid/circuit-grid.css
@@ -165,10 +165,15 @@
 
     /* Static badge positioning for mobile */
     .datamachine-day-badge {
-        position: static;
-        transform: none;
-        margin-bottom: 0.5rem;
+        /* Revert to absolute positioning for mobile to match desktop behavior */
+        position: absolute;
+        transform: translateY(-50%);
+        margin-bottom: 0;
         display: block;
+        opacity: 0; /* Will be made visible by JS */
+    }
+    
+    .datamachine-day-badge.positioned {
         opacity: 1;
     }
 }
