+++ inc/blocks/calendar/block.json
@@ -2,7 +2,7 @@
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 3,
 	"name": "datamachine-events/calendar",
-	"version": "0.2.2",
+	"version": "0.2.4",
 	"title": "Data Machine Events Calendar",
 	"category": "widgets",
 	"icon": "calendar-alt",
